@@ -4,6 +4,12 @@ import 'package:stacked/src/code_generation/stacked_router_annotations.dart';
 
 /// The annotation that defines a stacked application
 class StackedApp {
+  /// Supply to the StackedApp annotation a locatorName. The default name will be "locator".
+  final String locatorName;
+
+  /// Supply the StackedApp annotation with a locatorSetupName. The default name will be setupLocator
+  final String locatorSetupName;
+
   /// Defines all the routes that has to be generated for the onGenerateRoute function
   final List<StackedRoute> routes;
 
@@ -15,7 +21,9 @@ class StackedApp {
 
   const StackedApp({
     required this.routes,
-    this.dependencies,
     this.logger,
+    this.dependencies,
+    this.locatorName = 'locator',
+    this.locatorSetupName = 'setupLocator',
   });
 }
