@@ -1,4 +1,5 @@
 import 'package:firebase_crashlytics/firebase_crashlytics.dart';
+import 'package:flutter/material.dart';
 // ignore: implementation_imports
 import 'package:flutter/src/foundation/assertions.dart';
 import 'package:logger/logger.dart';
@@ -20,44 +21,79 @@ class CrashlyticsService {
   );
 
   void recordFlutterErrorToCrashlytics(FlutterErrorDetails details) {
-    _crashlyticsService.recordFlutterError(details);
+    try {
+      _crashlyticsService.recordFlutterError(details);
+    } catch (e) {
+      _catchOrThrow(e);
+    }
   }
 
   Future setUserIdToCrashlytics({String? id}) async {
-    if (id != null) await _crashlyticsService.setUserIdentifier(id);
+    try {
+      if (id != null) await _crashlyticsService.setUserIdentifier(id);
+    } catch (e) {
+      _catchOrThrow(e);
+    }
   }
 
   Future logToCrashlytics(
-      Level level, List<String> lines, StackTrace stacktrace,
-      {required bool logwarnings}) async {
-    if (level == Level.error || level == Level.wtf) {
-      await _crashlyticsService.recordError(
-        lines.join('\n'),
-        stacktrace,
-        printDetails: true,
-        fatal: true,
-      );
-    }
-    if (level == Level.warning && logwarnings) {
-      await _crashlyticsService.recordError(
-        lines.join('\n'),
-        stacktrace,
-        printDetails: true,
-      );
-    }
-    if (level == Level.info || level == Level.verbose || level == Level.debug) {
-      await _crashlyticsService.log(lines.join('\n'));
+    Level level,
+    List<String> lines,
+    StackTrace stacktrace, {
+    required bool logwarnings,
+  }) async {
+    try {
+      if (level == Level.error || level == Level.wtf) {
+        await _crashlyticsService.recordError(
+          lines.join('\n'),
+          stacktrace,
+          printDetails: true,
+          fatal: true,
+        );
+      }
+      if (level == Level.warning && logwarnings) {
+        await _crashlyticsService.recordError(
+          lines.join('\n'),
+          stacktrace,
+          printDetails: true,
+        );
+      }
+      if (level == Level.info ||
+          level == Level.verbose ||
+          level == Level.debug) {
+        await _crashlyticsService.log(lines.join('\n'));
+      }
+    } catch (exception) {
+      _catchOrThrow(exception);
     }
   }
 
   Future setCustomKeysToTrack(String key, dynamic value) async {
-    await _crashlyticsService.setCustomKey(key, value);
+    try {
+      await _crashlyticsService.setCustomKey(key, value);
+    } catch (e) {
+      _catchOrThrow(e);
+    }
   }
 
   // Be very careful when you excute this code it will crash the app
   // So, be sure to remove it after usage
   void crashApp() {
-    _crashlyticsService.crash();
+    try {
+      _crashlyticsService.crash();
+    } catch (e) {
+      _catchOrThrow(e);
+    }
+  }
+
+  void _catchOrThrow(dynamic exception) {
+    final _exceptionString = exception.toString();
+    final _isPluginConstantsException = _exceptionString
+        .contains("pluginConstants['isCrashlyticsCollectionEnabled']");
+
+    if (!_isPluginConstantsException) {
+      throw exception;
+    }
   }
 }
 
@@ -68,9 +104,14 @@ class CrashlyticsOutput extends LogOutput {
   @override
   void output(OutputEvent event) {
     try {
-      CrashlyticsService.getInstance().then((instance) => instance
-          .logToCrashlytics(event.level, event.lines, StackTrace.current,
-              logwarnings: logWarnings));
+      CrashlyticsService.getInstance().then((instance) {
+        return instance.logToCrashlytics(
+          event.level,
+          event.lines,
+          StackTrace.current,
+          logwarnings: logWarnings,
+        );
+      });
     } catch (e) {
       print('CRASHLYTICS FAILED: $e');
     }
