@@ -1,6 +1,8 @@
 import 'package:stacked_generator/src/generators/enums/dependency_type.dart';
 import 'package:recase/recase.dart';
 
+import 'stacked_locator_parameter_resolver.dart';
+
 class DependencyConfig {
   /// The import to use for the type of the service
   final String? import;
@@ -23,12 +25,18 @@ class DependencyConfig {
   /// The static function to use for resolving a singleton instance
   final String? resolveFunction;
 
+  final Set<String>? environments;
+
+  final Set<DependencyParamConfig>? params;
+
   DependencyConfig({
+    this.type,
+    this.params,
+    this.environments,
     required this.import,
     required this.className,
     this.abstractedImport,
     this.abstractedTypeClassName,
-    this.type,
     this.presolveFunction,
     this.resolveFunction,
   });
