@@ -10,9 +10,8 @@ import 'package:stacked_generator/src/generators/enums/dependency_type.dart';
 import 'package:stacked_generator/src/generators/getit/dependency_config.dart';
 import 'package:stacked_generator/src/generators/getit/stacked_locator_content_generator.dart';
 import 'package:stacked_generator/src/generators/getit/services_config.dart';
+import 'package:stacked_generator/src/generators/getit/stacked_locator_parameter_resolver.dart';
 import 'package:stacked_generator/utils.dart';
-
-const stackedRouteChecker = TypeChecker.fromRuntime(StackedRoute);
 
 class StackedLocatorGenerator extends GeneratorForAnnotation<StackedApp> {
   @override
@@ -23,6 +22,17 @@ class StackedLocatorGenerator extends GeneratorForAnnotation<StackedApp> {
   ) async {
     var libs = await buildStep.resolver.libraries.toList();
     var importResolver = ImportResolver(libs, element.source?.uri.path ?? '');
+
+    final String locatorName =
+        stackedApplication.peek('locatorName')!.stringValue;
+
+    final String locatorSetupName =
+        stackedApplication.peek('locatorSetupName')!.stringValue;
+
+    throwIf(
+      locatorName.isEmpty || locatorSetupName.isEmpty,
+      "Error: locatorName or locatorSetupName can not be a null or empty string",
+    );
 
     final servicesConfig = stackedApplication.peek('dependencies')?.listValue;
     if (servicesConfig != null) {
@@ -36,8 +46,11 @@ class StackedLocatorGenerator extends GeneratorForAnnotation<StackedApp> {
         services.add(serialisedServiceConfig);
       }
 
-      return StackedLocatorContentGenerator(ServicesConfig(services: services))
-          .generate();
+      return StackedLocatorContentGenerator(
+        servicesConfig: ServicesConfig(services: services),
+        locatorName: toLowerCamelCase(locatorName),
+        locatorSetupName: toLowerCamelCase(locatorSetupName),
+      ).generate();
     }
   }
 
@@ -65,6 +78,14 @@ class StackedLocatorGenerator extends GeneratorForAnnotation<StackedApp> {
       '🛑 ${toDisplayString(dependencyClassType)} is not a class element. All services should be classes. We don\'t register individual values for global access through the locator. Make sure the value provided as your service type is a class.',
     );
 
+    final Set<String>? environments = dependencyReader
+        .peek('environments')
+        ?.setValue
+        .map((e) => e.toStringValue())
+        .where((element) => element != null)
+        .toSet()
+        .cast<String>();
+
     // Get the import of the class type that's defined for the service
     final import = importResolver.resolve(classElement!);
 
@@ -81,6 +102,15 @@ class StackedLocatorGenerator extends GeneratorForAnnotation<StackedApp> {
 
     // NOTE: This can be used for actual dependency inject. We do service location instead.
     final constructor = classElement.unnamedConstructor;
+
+    final Set<DependencyParamConfig> clazzParams = {};
+    var params = constructor?.parameters;
+    if (params?.isNotEmpty == true && constructor != null) {
+      final paramResolver = DependencyParameterResolver(importResolver);
+      for (ParameterElement p in constructor.parameters) {
+        clazzParams.add(paramResolver.resolve(p));
+      }
+    }
 
     final serviceType = _getDependencyType(dependencyReader);
 
@@ -107,8 +137,10 @@ class StackedLocatorGenerator extends GeneratorForAnnotation<StackedApp> {
       import: import,
       abstractedImport: abstractedImport,
       type: serviceType,
+      params: clazzParams,
       presolveFunction: presolveFunction,
       resolveFunction: resolveFunction,
+      environments: environments,
     );
   }
 
@@ -124,6 +156,9 @@ class StackedLocatorGenerator extends GeneratorForAnnotation<StackedApp> {
     }
     if (serviceReader.instanceOf(TypeChecker.fromRuntime(Presolve))) {
       return DependencyType.PresolvedSingleton;
+    }
+    if (serviceReader.instanceOf(TypeChecker.fromRuntime(FactoryWithParam))) {
+      return DependencyType.FactoryWithParam;
     }
 
     return DependencyType.Singleton;
