@@ -4,39 +4,37 @@
 // InjectableConfigGenerator
 // **************************************************************************
 
-import 'package:get_it/get_it.dart';
-import 'package:injectable/injectable.dart';
-import 'package:stacked_services/stacked_services.dart';
+import 'package:get_it/get_it.dart' as _i1;
+import 'package:injectable/injectable.dart' as _i2;
+import 'package:stacked_services/stacked_services.dart' as _i3;
 
-import '../services/thirdparty_services_module.dart';
+import '../services/thirdparty_services_module.dart'
+    as _i4; // ignore_for_file: unnecessary_lambdas
 
-/// adds generated dependencies
-/// to the provided [GetIt] instance
-
-GetIt $initGetIt(
-  GetIt get, {
-  String environment,
-  EnvironmentFilter environmentFilter,
-}) {
-  final gh = GetItHelper(get, environment, environmentFilter);
+// ignore_for_file: lines_longer_than_80_chars
+/// initializes the registration of provided dependencies inside of [GetIt]
+_i1.GetIt $initGetIt(_i1.GetIt get,
+    {String environment, _i2.EnvironmentFilter environmentFilter}) {
+  final gh = _i2.GetItHelper(get, environment, environmentFilter);
   final thirdPartyServicesModule = _$ThirdPartyServicesModule();
-  gh.lazySingleton<BottomSheetService>(
+  gh.lazySingleton<_i3.BottomSheetService>(
       () => thirdPartyServicesModule.bottomSheetService);
-  gh.lazySingleton<DialogService>(() => thirdPartyServicesModule.dialogService);
-  gh.lazySingleton<NavigationService>(
+  gh.lazySingleton<_i3.DialogService>(
+      () => thirdPartyServicesModule.dialogService);
+  gh.lazySingleton<_i3.NavigationService>(
       () => thirdPartyServicesModule.navigationService);
-  gh.lazySingleton<SnackbarService>(
+  gh.lazySingleton<_i3.SnackbarService>(
       () => thirdPartyServicesModule.snackbarService);
   return get;
 }
 
-class _$ThirdPartyServicesModule extends ThirdPartyServicesModule {
+class _$ThirdPartyServicesModule extends _i4.ThirdPartyServicesModule {
   @override
-  BottomSheetService get bottomSheetService => BottomSheetService();
+  _i3.BottomSheetService get bottomSheetService => _i3.BottomSheetService();
   @override
-  DialogService get dialogService => DialogService();
+  _i3.DialogService get dialogService => _i3.DialogService();
   @override
-  NavigationService get navigationService => NavigationService();
+  _i3.NavigationService get navigationService => _i3.NavigationService();
   @override
-  SnackbarService get snackbarService => SnackbarService();
+  _i3.SnackbarService get snackbarService => _i3.SnackbarService();
 }
