@@ -8,7 +8,19 @@ void setupBottomSheetUi() {
 
   final builders = {
     BottomSheetType.FloatingBox: (context, sheetRequest, completer) =>
-        _FloatingBoxBottomSheet(request: sheetRequest, completer: completer)
+        _FloatingBoxBottomSheet(
+          request: sheetRequest,
+          completer: completer,
+        ),
+    BottomSheetType.Generic: (
+      context,
+      sheetRequest,
+      Function(SheetResponse<GenericBottomSheetResponse>) completer,
+    ) =>
+        GenericBottomSheet(
+          request: sheetRequest,
+          completer: completer,
+        ),
   };
 
   bottomSheetService.setCustomSheetBuilders(builders);
@@ -61,6 +73,94 @@ class _FloatingBoxBottomSheet extends StatelessWidget {
               ),
               TextButton(
                 onPressed: () => completer(SheetResponse(confirmed: true)),
+                child: Text(
+                  request.mainButtonTitle,
+                  style: TextStyle(color: Colors.white),
+                ),
+                style: ButtonStyle(
+                  backgroundColor: MaterialStateProperty.all<Color>(
+                    Theme.of(context).primaryColor,
+                  ),
+                ),
+              )
+            ],
+          )
+        ],
+      ),
+    );
+  }
+}
+
+class GenericBottomSheetRequest {
+  const GenericBottomSheetRequest({
+    this.message = 'GenericBottomSheetRequest',
+  });
+
+  final String message;
+}
+
+class GenericBottomSheetResponse {
+  const GenericBottomSheetResponse({
+    this.message = 'GenericBottomSheetResponse',
+  });
+
+  final String message;
+}
+
+class GenericBottomSheet extends StatelessWidget {
+  final SheetRequest request;
+  final Function(SheetResponse<GenericBottomSheetResponse>) completer;
+
+  const GenericBottomSheet({
+    Key key,
+    this.request,
+    this.completer,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      margin: EdgeInsets.all(25),
+      padding: EdgeInsets.all(25),
+      decoration: BoxDecoration(
+        color: Colors.white,
+        borderRadius: BorderRadius.circular(15),
+      ),
+      child: Column(
+        mainAxisSize: MainAxisSize.min,
+        children: [
+          Text(
+            request.title ?? 'Generic Bottom Sheet',
+            style: TextStyle(
+              fontSize: 20,
+              fontWeight: FontWeight.bold,
+              color: Colors.grey[900],
+            ),
+          ),
+          SizedBox(height: 10),
+          Text(
+            request.description,
+            style: TextStyle(color: Colors.grey),
+          ),
+          Row(
+            mainAxisSize: MainAxisSize.max,
+            mainAxisAlignment: MainAxisAlignment.spaceBetween,
+            children: [
+              MaterialButton(
+                onPressed: () => completer(SheetResponse(
+                  confirmed: true,
+                  data: GenericBottomSheetResponse(message: 'SecondaryButton'),
+                )),
+                child: Text(
+                  request.secondaryButtonTitle,
+                  style: TextStyle(color: Theme.of(context).primaryColor),
+                ),
+              ),
+              TextButton(
+                onPressed: () => completer(SheetResponse(
+                  confirmed: true,
+                  data: GenericBottomSheetResponse(message: 'MainButton'),
+                )),
                 child: Text(
                   request.mainButtonTitle,
                   style: TextStyle(color: Colors.white),
