@@ -8,7 +8,13 @@ void setupDialogUi() {
 
   final builders = {
     DialogType.Basic: (context, sheetRequest, completer) =>
-        _BasicDialog(request: sheetRequest, completer: completer)
+        _BasicDialog(request: sheetRequest, completer: completer),
+    DialogType.Generic: (context, sheetRequest,
+            Function(DialogResponse<GenericDialogResponse>) completer) =>
+        _GenericDialog(
+          request: sheetRequest,
+          completer: completer,
+        ),
   };
 
   dialogService.registerCustomDialogBuilders(builders);
@@ -51,6 +57,86 @@ class _BasicDialog extends StatelessWidget {
             ),
             GestureDetector(
               onTap: () => completer(DialogResponse()),
+              child: Container(
+                child: request.showIconInMainButton
+                    ? Icon(Icons.check_circle)
+                    : Text(request.mainButtonTitle),
+                alignment: Alignment.center,
+                padding: const EdgeInsets.symmetric(vertical: 10),
+                width: double.infinity,
+                decoration: BoxDecoration(
+                  color: Colors.redAccent,
+                  borderRadius: BorderRadius.circular(5),
+                ),
+              ),
+            )
+          ],
+        ),
+      ),
+    );
+  }
+}
+
+class GenericDialogRequest {
+  const GenericDialogRequest({
+    this.message = 'Hello World',
+  });
+
+  final String message;
+}
+
+class GenericDialogResponse {
+  const GenericDialogResponse({
+    this.message = 'Hello World',
+  });
+
+  final String message;
+}
+
+class _GenericDialog extends StatelessWidget {
+  final DialogRequest request;
+  final Function(DialogResponse<GenericDialogResponse>) completer;
+
+  const _GenericDialog({
+    Key key,
+    this.request,
+    this.completer,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Dialog(
+      child: Container(
+        padding: const EdgeInsets.all(20),
+        decoration: BoxDecoration(
+          color: Colors.white,
+          borderRadius: BorderRadius.circular(10),
+        ),
+        child: Column(
+          crossAxisAlignment: CrossAxisAlignment.center,
+          mainAxisSize: MainAxisSize.min,
+          children: <Widget>[
+            Text(
+              request.title ?? 'Generic Dialog',
+              style: TextStyle(fontWeight: FontWeight.bold, fontSize: 23),
+            ),
+            SizedBox(
+              height: 10,
+            ),
+            Text(
+              request.description,
+              style: TextStyle(
+                fontSize: 18,
+              ),
+              textAlign: TextAlign.center,
+            ),
+            SizedBox(
+              height: 20,
+            ),
+            GestureDetector(
+              onTap: () => completer(
+                DialogResponse(data: GenericDialogResponse()),
+              ),
               child: Container(
                 child: request.showIconInMainButton
                     ? Icon(Icons.check_circle)
