@@ -1,18 +1,22 @@
 import 'package:flutter/material.dart';
+import 'package:mockito/annotations.dart';
 import 'package:mockito/mockito.dart';
 import 'package:stacked_themes/src/locator_setup.dart';
+import 'package:stacked_themes/src/services/platform_service.dart';
 import 'package:stacked_themes/src/services/shared_preferences_service.dart';
 import 'package:stacked_themes/src/services/statusbar_service.dart';
 
-class SharedPreferencesServiceMock extends Mock
-    implements SharedPreferencesService {}
+import 'test_setup.mocks.dart';
 
-class StatusBarServiceMock extends Mock implements StatusBarService {}
-
+@GenerateMocks([], customMocks: [
+  MockSpec<SharedPreferencesService>(returnNullOnMissingStub: true),
+  MockSpec<StatusBarService>(returnNullOnMissingStub: true),
+  MockSpec<PlatformService>(returnNullOnMissingStub: true)
+])
 SharedPreferencesService getAndRegisterSharedPreferencesServiceMock(
     {int? themeIndex, ThemeMode? userThemeMode}) {
   _removeRegistrationIfExists<SharedPreferencesService>();
-  var service = SharedPreferencesServiceMock();
+  var service = MockSharedPreferencesService();
 
   when(service.themeIndex).thenReturn(themeIndex);
   when(service.userThemeMode).thenReturn(userThemeMode);
@@ -23,8 +27,18 @@ SharedPreferencesService getAndRegisterSharedPreferencesServiceMock(
 
 StatusBarService getAndRegisterStatusBarServiceMock() {
   _removeRegistrationIfExists<StatusBarService>();
-  var service = StatusBarServiceMock();
+  var service = MockStatusBarService();
   locator.registerSingleton<StatusBarService>(service);
+  return service;
+}
+
+PlatformService getAndRegisterPlatformService() {
+  _removeRegistrationIfExists<PlatformService>();
+  final service = MockPlatformService();
+
+  when(service.isMobilePlatform).thenReturn(true);
+
+  locator.registerSingleton<PlatformService>(service);
   return service;
 }
 
@@ -39,9 +53,11 @@ void _removeRegistrationIfExists<T extends Object>() {
 void registerServices() {
   getAndRegisterSharedPreferencesServiceMock();
   getAndRegisterStatusBarServiceMock();
+  getAndRegisterPlatformService();
 }
 
 void unregisterServices() {
   locator.unregister<SharedPreferencesService>();
   locator.unregister<StatusBarService>();
+  locator.unregister<PlatformService>();
 }
